refactor(mito): Move tests to a separate file
